integrated functionality into main script
